requirements.txt
datasets
torch
torchvision
transformers
numpy
tiktoken
tqdm
fastapi
pandas
matplotlib
pip
torchaudio  
accelerate
python-docx 
docx2txt